Moved ScreenshotCapturer and ScreenshotOnFailure (currently private[scalatest])  to package org.scalatestplus.selenium.
