+++ scalatest-test/src/test/scala/org/scalatestplus/selenium/JettySpec.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.selenium
+package org.scalatestplus.selenium
 
 import org.scalatest._
 import org.eclipse.jetty.server.Server
