+++ scalatest/src/main/scala/org/scalatestplus/selenium/ScreenshotCapturer.scala
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest
+package org.scalatestplus.selenium
 
 
 /**
  * Trait containing one abstract method that can capture a screenshot and save it as a file to a given directory.
  */
-private[scalatest] trait ScreenshotCapturer {
+private[selenium] trait ScreenshotCapturer {
 
   /**
    * Captures a screenshot and saves it as a file in the specified directory.
