+++ scalatest/src/main/scala/org/scalatestplus/selenium/WebBrowser.scala
@@ -45,7 +45,6 @@ import org.openqa.selenium.support.ui.Select
 import org.scalatest.exceptions.TestFailedException
 import org.scalatest.exceptions.StackDepthException
 import org.openqa.selenium.JavascriptExecutor
-import org.scalatest.ScalaTestInternals.ScreenshotCapturer
 import org.scalatest.time.Nanosecond
 import org.scalatest.Resources
 import org.scalactic.source
